feat: add charts page
