+++ src/App.less
@@ -43,6 +43,31 @@ body {
   position: relative;
   padding: 1rem;
   z-index: 2;
+
+  .ant-menu-horizontal {
+    border-bottom-color: transparent;
+    background-color: transparent;
+    line-height: inherit;
+    font-size: 16px;
+    margin: 0 10px;
+
+    .ant-menu-item {
+      margin: 0 10px;
+      color: lightgrey;
+      height: 35px;
+      line-height: 35px;
+      border-width: 0px !important;
+    }
+
+    .ant-menu-item:hover {
+      color: white;
+      border-width: 0px !important;
+    }
+    
+    .ant-menu-item-selected {
+      font-weight: bold;
+    }
+  }
 }
 
 .App-Bar-left {
