+++ src/components/appBar.tsx
@@ -1,0 +1,72 @@
+import React from "react";
+import { Button, Menu, Popover } from "antd";
+import { PoolAccounts } from "./pool/view";
+import { useWallet } from "../utils/wallet";
+import { AccountInfo } from "./accountInfo";
+import { useLocation } from 'react-router-dom'
+
+export const AppBar = (props: { left?: JSX.Element, right?: JSX.Element }) => {
+  const { connected, wallet } = useWallet();
+  const location = useLocation();
+
+  const TopBar = (
+    <div className="App-Bar">
+      <div className="App-Bar-left">
+        <div className="App-logo" />
+        <Menu mode="horizontal" selectedKeys={[location.pathname]}>
+          <Menu.Item key="/">
+            <a href={"/#"}>Swap</a>
+          </Menu.Item>
+          <Menu.Item key="/info">
+            <a href={"/#/info"}>Charts</a>
+          </Menu.Item>
+          <Menu.Item key="trade">
+            <a
+              href={"https://dex.projectserum.com"}
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              Trade
+              <sup>↗</sup>
+            </a>
+          </Menu.Item>
+        </Menu>
+        {props.left}
+      </div>
+      <div className="App-Bar-right">
+        <AccountInfo />
+        {connected && (
+          <Popover
+            placement="bottomRight"
+            content={<PoolAccounts />}
+            trigger="click"
+          >
+            <Button type="text">My Pools</Button>
+          </Popover>
+        )}
+        <div>
+          {!connected && (
+            <Button
+              type="text"
+              size="large"
+              onClick={connected ? wallet.disconnect : wallet.connect}
+              style={{ color: "#2abdd2" }}
+            >
+              Connect
+            </Button>
+          )}
+          {connected && (
+            <Popover
+              placement="bottomRight"
+              title="Wallet public key"
+              trigger="click"
+            ></Popover>
+          )}
+        </div>
+        {props.right}
+      </div>
+    </div>
+  );
+
+  return TopBar;
+};
