+++ src/components/exchange.tsx
@@ -2,14 +2,11 @@ import React, { useState } from "react";
 import { Button, Card, Popover } from "antd";
 import { TradeEntry } from "./trade";
 import { AddToLiquidity } from "./pool/add";
-import { PoolAccounts } from "./pool/view";
-import { useWallet } from "../utils/wallet";
-import { AccountInfo } from "./accountInfo";
 import { Settings } from "./settings";
 import { SettingOutlined } from "@ant-design/icons";
+import { AppBar } from './appBar';
 
 export const ExchangeView = (props: {}) => {
-  const { connected, wallet } = useWallet();
   const tabStyle: React.CSSProperties = { width: 120 };
   const tabList = [
     {
@@ -30,72 +27,23 @@ export const ExchangeView = (props: {}) => {
 
   const [activeTab, setActiveTab] = useState(tabList[0].key);
 
-  const TopBar = (
-    <div className="App-Bar">
-      <div className="App-Bar-left">
-        <div className="App-logo" />
-      </div>
-      <div className="App-Bar-right">
-        <Button type="text" size="large" style={{ color: "#2abdd2" }}>
-          <a
-            href={"https://dex.projectserum.com"}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Trade
-          </a>
-        </Button>
-        <AccountInfo />
-        {connected && (
-          <Popover
-            placement="bottomRight"
-            content={<PoolAccounts />}
-            trigger="click"
-          >
-            <Button type="text">My Pools</Button>
-          </Popover>
-        )}
-        <div>
-          {!connected && (
-            <Button
-              type="text"
-              size="large"
-              onClick={connected ? wallet.disconnect : wallet.connect}
-              style={{ color: "#2abdd2" }}
-            >
-              Connect
-            </Button>
-          )}
-          {connected && (
-            <Popover
-              placement="bottomRight"
-              title="Wallet public key"
-              trigger="click"
-            ></Popover>
-          )}
-        </div>
-        {
-          <Popover
-            placement="topRight"
-            title="Settings"
-            content={<Settings />}
-            trigger="click"
-          >
-            <Button
-              shape="circle"
-              size="large"
-              type="text"
-              icon={<SettingOutlined />}
-            />
-          </Popover>
-        }
-      </div>
-    </div>
-  );
-
   return (
     <>
-      {TopBar}
+      <AppBar right={
+        <Popover
+          placement="topRight"
+          title="Settings"
+          content={<Settings />}
+          trigger="click"
+        >
+          <Button
+            shape="circle"
+            size="large"
+            type="text"
+            icon={<SettingOutlined />}
+          />
+        </Popover>
+      } />
       <Card
         className="exchange-card"
         headStyle={{ padding: 0 }}
