+++ src/routes.tsx
@@ -1,6 +1,7 @@
 import { HashRouter, Route } from "react-router-dom";
 import React from "react";
 import { ExchangeView } from "./components/exchange";
+import { ChartsView } from "./components/charts";
 
 export function Routes() {
   // TODO: add simple view for sharing ...
@@ -8,6 +9,7 @@ export function Routes() {
     <>
       <HashRouter basename={"/"}>
         <Route exact path="/" component={ExchangeView} />
+        <Route exact path="/info" component={ChartsView} />
       </HashRouter>
     </>
   );
